requirements.txt
click>=7.0
h5py>=2.9
NREL-PySAM==1.2.1
NREL-rex>=0.1.1
numpy>=1.16
pandas>=0.25
psutil>=5.6
scipy>=1.3
